Background cosmology

Description

Cosmojuly provides a set of functions for the background cosmology.

Structures

Main.Cosmojuly.BackgroundCosmo.BkgCosmology — Type

BkgCosmology{T<:Real}

Abstract type: generic background cosmology

Main.Cosmojuly.BackgroundCosmo.FLRW — Type

FLRW{T<:Real} <: BkgCosmology{T}

Abstract type: Generic background cosmology of type FLRW

Main.Cosmojuly.BackgroundCosmo.FlatFLRW — Type

FlatFLRW{T<:Real} <: FLRW{T}

Defines a flat FLRW cosmology

Main.Cosmojuly.BackgroundCosmo.FlatFLRW — Method

FlatFLRW(h, Ω_χ0, Ω_b0; T0_CMB_K = 2.72548, Neff = 3.04)

Creates a FlatFLRW instance

Arguments

h::Real : hubble parameter (dimensionless)

Ω_χ0::Real: cold dark matter abundance today (dimensionless)

Ω_b0::Real: baryon abundance today (dimensionless)

T0_CMB_K::Real: temperature of the CMB today (in Kelvin)

Neff::Real: effective number of neutrinos (dimensionless)

Constants

The module provides some predefined background cosmologies

Main.Cosmojuly.BackgroundCosmo.edsPlanck18_bkg — Constant

edsplanck18_bkg::FlatFLRW  = FlatFLRW(0.6736, 0.3, 0)

Flat FLRW cosmology with Planck18 parameters and no baryons

Main.Cosmojuly.BackgroundCosmo.planck18_bkg — Constant

planck18_bkg::FlatFLRW = FlatFLRW(0.6736, 0.26447, 0.04930)

Flat FLRW cosmology with Planck18 parameters

Functions

Main.Cosmojuly.BackgroundCosmo.convert_cosmo — Method

Convert cosmo object attributes to another type

Main.Cosmojuly.BackgroundCosmo.growth_factor — Function

growth_factor(z, cosmo)

Exact growth factor in a matter-Λ Universe computed from an integral
Carroll et al. 1992 (Mo and White p. 172)
Corresponds to D1(a=1) with the definition of Dodelson 2003

Arguments

z: redshift

cosmo: background cosmology (default Planck18)

Main.Cosmojuly.BackgroundCosmo.growth_factor_Carroll — Function

growth_factor_Carroll(z, cosmo)

Approximate growth factor in a matter-Λ Universe (faster than growth_factor)
Carroll et al. 1992 (Mo and White p. 172)
Corresponds to D1(a=1) with the definition of Dodelson 2003

Arguments

z: redshift

cosmo: background cosmology (default Planck18)

Main.Cosmojuly.BackgroundCosmo.hubble_constant — Function

Hubble constant H0 (in km/s/Mpc) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.hubble_evolution — Function

Hubble evolution (no dimension) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.hubble_rate — Function

Hubble rate H(z) (in km/s/Mpc) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.temperature_CMB_K — Function

CMB temperature (in K) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.Ω — Method

Ω(T, z = 0, cosmo = planck18_bkg) where {T<:Species}

Abundance of species T at redshift z and for the background cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_b_Msun_Mpc3 — Function

Baryon density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_c_Msun_Mpc3 — Function

Critical density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_m_Msun_Mpc3 — Function

Matter density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_r_Msun_Mpc3 — Function

Radiation density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_Λ_Msun_Mpc3 — Function

Cosmological constant density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_γ_Msun_Mpc3 — Function

Photon density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_ν_Msun_Mpc3 — Function

Neutrino density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo

Main.Cosmojuly.BackgroundCosmo.ρ_χ_Msun_Mpc3 — Function

Cold dark matter density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology cosmo