The FSL model

Description

Write here a short description of the model with the main equations

Functions

Main.Cosmojuly.FSLModel.subhalo_mass_function_template — Method

subhalo_mass_function_template(x, γ1, α1, γ2, α2, β, ζ)

Template function for the subhalo mass function:

$m_Δ^{\rm host} \frac{\partial N(m_Δ^{\rm sub}, z=0)}{\partial m_Δ^{\rm sub}} = \left(\gamma_1 x^{-\alpha_1} + \gamma_2 x^{-\alpha_2}\right) e^{-\beta x^\zeta}$

The first argument, x::Real, is the ratio of the subhalo over the host mass $m_Δ^{\rm sub} / m_Δ^{\rm host}.$

Main.Cosmojuly.FSLModel.mass_function_merger_tree — Method

mass_function_merger_tree(mΔ_sub, mΔ_host)

Example of subhalo mass function fitted on merger tree results (Facchinetti et al., in prep.)

Arguments

mΔ_sub::Real : subhalo virial mass (in Msun)

mΔ_host::Real: host virial mass (in Msun)

Returns

$\frac{\partial N(m_Δ^{\rm sub}, z=0)}{\partial m_Δ^{\rm sub}}$