Public API

Structure

DensityFlows.Flow — Type

Normalizing flow

DensityFlows.FlowElement — Type

Building blocks of the flow

DensityFlows.AffineCouplingElement — Type

AffineCouplingElement <: FlowElement

DensityFlows.RNVPCouplingLayer — Type

RNVPCouplingLayer <: AffineCouplingLayer

Can be called as a functor f::RNVPCouplingLayer(z, θ=nothing) equivalent to forward(f, z, θ)

The hierarchy of types can be visualised from the following outputs

julia> using DensityFlows
julia> AffineCouplingElement <: FlowElementtrue
julia> AffineCouplingLayer <: AffineCouplingElementtrue
julia> RNVPCouplingLayer <: AffineCouplingLayertrue

Axes

Axes define and manipulate dimensions on which the AffineCouplingElement operates. Some dimensions are left unchanged while the others undergo an affine transformation.

DensityFlows.AffineCouplingAxes — Type

AffineCouplingAxes(d, n=0; kws...)

Create axes for AffineCouplingLayer.

Arguments

d::Int: dimension of the flow.

n::Int: number of conditions / parameters (default is 0).

j::Int: dimension cut (default is d÷2).

reverse::Bool: (default is false).

The dimension cut j specifies which dimensions are not modified by the layer. If reverse is false the layer acts as the identity on dimensions (1, j). If reverse is true the layer acts as the identity on dimensions (j+1, d).

Axes can be manipulated with the following functions.

Base.reverse — Function

Base.reverse(axes)

Swap the dimensions that are left unchanged by the layer. See also AffineCouplingAxes.

Coupling elements

DensityFlows.AffineCouplingLayer — Type

AffineCouplingLayer(axes; kws...)
AffineCouplingLayer(d, n=0; kws...)

Create an AffineCouplingLayer with NN models s and t.

The layer can be represented as a function f such that on dimensions where it does not act like the identity it returns

\[ f(x) = x * \exp(s) + t \quad {\rm if~forward}\]

and

\[ f^{-1}(z) = \exp(-s) * (z-t) \quad {\rm if~backward} \, .\]

Arguments

axes::AffineCouplingAxes.

d::Int: dimension of the flow.

n::Int: number of conditions / parameters (default is 0).

j::Int: dimension cut (default is d÷2).

reverse::Bool: (default is false).

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

If axes is provided, then d, n, j and reverse should not be passed as arguments as they would be redondant.

The dimension cut j specifies which dimensions are not modified by the layer. If reverse is false the layer acts as the identity on dimensions (1, j). If reverse is true the layer acts as the identity on dimensions (j+1, d).

See also AffineCouplingAxes.

DensityFlows.AffineCouplingBlock — Type

AffineCouplingBlock(axes; kws...)

Create an block of two AffineCouplingLayer with opposite axes.

Opposite axes here means that one is set with reverse = true and the other with reverse = false.

Arguments

axes::AffineCouplingAxes.

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

Chains

DensityFlows.AffineCouplingChain — Type

AffineCouplingChain(n_couplings, axes, U; kws...)
AffineCouplingChain(xs...)

Create an chain of AffineCouplingLayer or AffineCouplingBlock.

Can either create a chain of n_couplings similar layers or blocks by passing n_couplings and axes or instantiate a chain from pre-existings layers or blocks passed as xs.

Arguments

n_couplings::Int: number of couplings.

axes::AffineCouplingAxes.

U:Type: type of struct in the chain, can be AffineCouplingLayer or AffineCouplingBlock (default is AffineCouplingBlock).

Keywords arguments kws... are passed to the constructor of AffineCouplingLayer or AffineCouplingBlock.

Evaluation

DensityFlows.backward — Function

backward(f, x, θ=nothing)

Return $f^{-1}(x \,|\, \theta)$ and $J_{f^{-1}}(x \,| \,\theta)$ where $\theta$ is an array of parameters.

Arguments

x::AbstractArray{T, N}: arguments to pass to the flow element.

θ::Union{AbstractArray{T, N}, Nothing}: parameters / conditions (default is nothing).

DensityFlows.forward — Function

forward(f, x, θ=nothing)

Return $f(z \,| \,\theta)$ and $J_{f}(z \,| \,\theta)$ where $\theta$ is an array of parameters.

Arguments

z::AbstractArray{T, N}: arguments to pass to the flow element.

θ::Union{AbstractArray{T, N}, Nothing}: parameters / conditions (default is nothing).

Save and loads

DensityFlows.save — Function

save(directory, element; kws...)

Recursively save the FlowElement element's weights in directory.

Setting erase = true force deleted any existing directory with the same name.

DensityFlows.load — Function

load(directory)

Load any FlowElement saved in directory.

Data

DensityFlows.DataArrays — Type

DataArrays(x, θ = nothing, f_training = 0.9, f_validation = 0.1, rng = Random.default_rng())

Normalised and partitioned data to feed the neural network.

x must be of size (d, ...) where d is the number of physical dimensions. θ must be of size (n, ...) where n is the number of parameters and every other array dimensions in place of ... should match that of x.

Warning

Data is partitioned along the second axis only. It is thus necessary to make sure that size(x, 2) is large enough by swapping some of the dimensions if necessary.

See also DataPartition and MetaData.

DensityFlows.MetaData — Type

MetaData(hash, x_min, x_max, θ_min, θ_max)

Metadata containing an identification hash value and the boundaries of the data and parameters arrays x and θ.

DensityFlows.DataPartition — Type

DataPartition(n, f_training = 0.9, f_validation = 0.1, rng = Random.default_rng())

Random partition of the data.

The data is devided into a fraction f_training of training data and a fraction f_validation of validation data. If f_training + f_validation < 1 the rest is kept as testing data.

DensityFlows.data_x_min — Function

Minimum value of the input data.

DensityFlows.data_x_max — Function

Maximum value of the input data.

DensityFlows.data_θ_min — Function

Minimin value of the input parameters.

DensityFlows.data_θ_max — Function

Maximum value of the input parameters.

DensityFlows.normalize_input — Function

normalise_input(x, x_min, x_max)

Normalize input between $[0, 1]$.

\[ y = \frac{x - x_{\rm min}}{x_{\rm max} - x_{\rm min}}\]

See also normalize_input! and resize_output.

DensityFlows.normalize_input! — Function

normalise_input!(x, x_min, x_max)

Normalize input between $[0, 1]$ in place.

See also normalize_input and resize_output!.

DensityFlows.resize_output — Function

normalise_input(x, x_min, x_max)

Resize the output between $[x_{\rm min}, x_{\rm max}]$.

\[ x = (x_{\rm max} - x_{\rm min}) y + x_{\rm min}\]

See also resize_output! and normalize_input.

DensityFlows.resize_output! — Function

normalise_input!(x, x_min, x_max)

Resize the output between $[x_{\rm min}, x_{\rm max}]$ in place.

See also resize_output and normalize_input!.