Couplings

Axes

Axes define and manipulate dimensions on which the CouplingLayer operates. Some dimensions are left unchanged while the others undergo an affine transformation.

DensityFlows.CouplingAxes — Type

CouplingAxes(d, mask; kws )
CouplingAxes(d, j=d÷2; kws...)
CouplingAxes(data, mask)
CouplingAxes(data, j=d÷2; reverse)

Create axes for CouplingLayer.

Arguments

d::Int: dimension of the flow.

j::Int: dimension cut (default is d÷2).

mask::AbstractVector{Int}: dimensions that are affected by the coupling.

n::Int: number of conditions / parameters (default is 0).

reverse::Bool: (default is false).

The dimension cut j specifies which dimensions are not modified by the layer. If reverse is false the layer acts as the identity on dimensions (1, j). If reverse is true the layer acts as the identity on dimensions (j+1, d).

Axes can be manipulated with the following functions.

Base.reverse — Function

Base.reverse(axes)

Swap the dimensions that are left unchanged by the layer. See also CouplingAxes.

Coupling elements

DensityFlows.CouplingLayer — Type

CouplingLayer([T=RNVPCouplingLayer, ] axes; kws...)
CouplingLayer([T=RNVPCouplingLayer, ] d, j = d ÷ 2; n=0, reverse=false, kws...)
CouplingLayer([T=RNVPCouplingLayer, ] d, mask; n=0, kws...)
CouplingLayer([T=RNVPCouplingLayer, ] data, mask; kws...)
CouplingLayer([T=RNVPCouplingLayer, ] data, j = d ÷ 2; reverse=false, kws...)

CouplingLayer(t_net, axes)
CouplingLayer(s_net, t_net, axes)

Create an CouplingLayer with NN models s and t.

The layer can be represented as a function f such that on dimensions where it does not act like the identity it returns

\[ f(x) = x * \exp(s) + t \quad {\rm if~forward}\]

and

\[ f^{-1}(z) = \exp(-s) * (z-t) \quad {\rm if~backward} \, .\]

By default s and t are built with Dense neural networks.

Arguments

axes::CouplingAxes.

d::Int: dimension of the flow.

j::Int: dimension cut (default is d÷2).

mask::Vector{Int}: dimensions that are affected by the coupling.

data::DataArrays: input data arrays

Keywords arguments

n::Int: number of conditions / parameters (default is 0).

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

σ::Function: activation function (default is Flux.relu).

bias::Bool: activate bias (default is true).

Example

julia> @summary CouplingLayer(3, [1, 3], n=2, hidden_dim=10, n_sublayers_s=1, σ=Flux.tanh)
RNVPCouplingLayer | s_net > [3, 10, 2] (62 parameters)
                  | t_net > [3, 10, 10, 2] (172 parameters)
                  | axes  > (d,n)=(3,2); identity=(2), transformed=(1,3)

See also CouplingAxes.

DensityFlows.CouplingBlock — Type

CouplingBlock(layer_1, layer_2)
CouplingBlock([T=RNVPCouplingLayer, ] first_axes; kws...)
CouplingBlock([T=RNVPCouplingLayer, ] d, j = d ÷ 2; n=0, reverse=false, kws...)
CouplingBlock([T=RNVPCouplingLayer, ] d, mask; n=0, kws...)

Create an block of two CouplingLayer with opposite / complementary axes.

Arguments

first_axes::CouplingAxes: axes of the first layer.

d::Int: dimension of the flow.

j::Int: dimension cut (default is d÷2).

mask::Vector{Int}: dimensions that are affected by the coupling.

Keyword arguments

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

σ::Function: activation function (default is Flux.relu).

bias::Bool: activate bias (default is true).).

Example

julia> @summary CouplingBlock(3, [1, 3], n=2, hidden_dim=10, n_sublayers_s=1, σ=Flux.tanh)
RNVPCouplingLayer | s_net > [3, 10, 2] (62 parameters)
                  | t_net > [3, 10, 10, 2] (172 parameters)
                  | axes  > (d,n)=(3,2); identity=(2), transformed=(1,3)
RNVPCouplingLayer | s_net > [4, 10, 1] (61 parameters)
                  | t_net > [4, 10, 10, 1] (171 parameters)
                  | axes  > (d,n)=(3,2); identity=(1,3), transformed=(2)

DensityFlows.RNVPCouplingLayer — Type

RNVPCouplingLayer(s_net, t_net, axes)

Structure of a Real-Non-Volume-Preserving affine coupling layer.

Contains

Specific functions

DensityFlows.RNVP_backward — Function

RNVP_backward(s, t, u, axis_id, axis_af)

Return z = (u-t)*exp(s), ln(det|J|) = sum(s).

DensityFlows.NICE_backward — Function

NICE_backward(t, u, axis_id, axis_af)

Return z = (u-t), zeros(...).