Data

Structures

DensityFlows.DataArrays — Type

DataArrays(x, θ = dflt_θ(x); f_training = 0.9, f_validation = 0.1, rng = Random.default_rng())

Normalised and partitioned data to feed the neural network.

x must be of size (d, ...) where d is the number of physical dimensions. θ must be of size (n, ...)where n is the number of parameters and every other array dimensions in place of ... should match that of x.

Warning

Data is partitioned along the second axis only. It is thus necessary to make sure that size(x, 2) is large enough by swapping some of the dimensions if necessary.

See also DataPartition and MetaData.

DensityFlows.MetaData — Type

MetaData(hash, θ_min, θ_max)

Metadata containing an identification hash value and the boundaries of the parameters array θ.

DensityFlows.DataPartition — Type

DataPartition(n, f_training = 0.9, f_validation = 0.1, rng = Random.default_rng())

Random partition of the data.

The data is devided into a fraction f_training of training data and a fraction f_validation of validation data. If f_training + f_validation < 1 the rest is kept as testing data.

Functions

DensityFlows.dflt_θ — Function

dflt_θ([T = Float32,] dims::Tuple)
dflt_θ([T = Float32,] dims...)
dflt_θ(x::AbstractArray)

Default value of the parameters, setting the first dimension to size 0.

Arguments are the same than zeros or ones. If an array x is passed returns an array of the same dimensions than x with the first dimension set to size 0.

Examples

julia> dflt_θ(2, 3)
0×2×3 Array{Float32, 3}

julia> dflt_θ(ones(2, 4, 5))
0×4×5 Array{Float64, 3}

DensityFlows.minimum_θ — Function

Minimin value of the input parameters.

DensityFlows.maximum_θ — Function

Maximum value of the input parameters.

DensityFlows.normalize_input — Function

normalise_input(x, x_min, x_max)

Normalize input between $[-1, 1]$.

\[ y = \frac{x - x_{\rm min}}{x_{\rm max} - x_{\rm min}}\]

See also normalize_input! and resize_output.

DensityFlows.normalize_input! — Function

normalise_input!(x, x_min, x_max)

Normalize input between $[-1, 1]$ in place.

See also normalize_input and resize_output!.

DensityFlows.resize_output — Function

normalise_input(x, x_min, x_max)

Resize the output between $[x_{\rm min}, x_{\rm max}]$.

\[ x = (x_{\rm max} - x_{\rm min}) y + x_{\rm min}\]

See also resize_output! and normalize_input.

DensityFlows.resize_output! — Function

normalise_input!(x, x_min, x_max)

Resize the output between $[x_{\rm min}, x_{\rm max}]$ in place.

See also resize_output and normalize_input!.