Flow

DensityFlows.Flow — Type

Flow([base, ] model, data)

Create a Flow from a model FlowChain and for a specific data.

The data must be passed as a DataArrays. A spceific base distribution can be given, default is multivariate gaussian.

Chains

DensityFlows.FlowChain — Type

FlowChain(elements::Tuple)
FlowChain(elements...)
FlowChain([T = CouplingBlock, ], n, args...; kwars... )

Instanciate a chain of flow elements from a Tuple.

Possible to directly pass the elements of a chain, or construct a chain of n identical blocks of type T, with args... and kws... passed to the constructor of T.

DensityFlows.concatenate — Function

concatenate(x::FlowChain...)
concatenate(x::FlowChain, y::FlowElement...)
concatenate(x::Union{Tuple, FlowElement}, y::FlowChain...)

Make one FlowChain from multiple chains or adding flow elements.

Density evaluation

StatsBase.sample — Function

sample([rng=default_rng(), ] flow, dims [, θ = dflt_θ(T, dims)] )

Sample the flow distribution.

Give a sample of the flow distribution of size given by dims which can be an Integer, or NTuple of integers. The default random sampler is Random.default_rng() but any sampler of type Random.AbstractRNG can be used.

If the flow is conditional, θ can be passed as an array of size (n, dims...) where n is the number of parameters if θ different for at least two sampled points. Otherwise θ can be given as a NTuple{n, T}. In that case all points are drawn with the same θ parameter.

Example

# return a sample of size (20, 10)
sample(flow, (20, 10))

# return a sample of size (20, 10) at 
# parameter (1f0, 2f0) if flow is conditional
sample(flow, (20, 10), (1f0, 2f0))

Distributions.logpdf — Function

logpdf(flow, x [, θ = dflt_θ(x)])

Natural logarithm of the probability density function given by the flow.

Argument x can be given as an array of size (d, dims...) or as a Tuple of d vectors of each length. In the latter case, return the logpdf on a grid of values defined by these vectors.

If the flow is conditional, θ can be passed as an array of size (n, dims...) where n is the number of parameters if θ different for at least two sampled points. Otherwise θ can be given as a NTuple{n, T}.

Warning

If x is given as a Tuple of vectors, θ must be passed as a NTuple{n, T}.

Example

# Give arrays in the training range of the flow
x = range(1f0, 10f0, 40)
y = range(-2.5f0, 11f0, 10)
z = range(0.1f0, 2f0, 30)

# For a given trained Flow 'flow' on 3 dimensions
res = logpdf(flow, (x, y, z)) # if unconditional
res = logpdf(flow, (x, y, z), (1f0, 2f0)) # if 2 conditions

# using contour in Plots
contour(x, y, res[:, :, 1]')
contour(x, z, res[:, 4, :]')

See also pdf.

Distributions.pdf — Function

pdf(flow, x, [, θ = dflt_θ(x)])

Probability density function given by the flow.

See also logpdf.