Overview

Structure

DensityFlows.FlowElement — Type

Building blocks of the flow

DensityFlows.summarize — Function

summarize(element)

Print a summary for the FlowElement element.

The hierarchy of types can be visualised from the following outputs

julia> using DensityFlows
julia> CouplingLayer <: FlowElementtrue
julia> RNVPCouplingLayer <: CouplingLayertrue

Evaluation

DensityFlows.backward — Function

backward(f, x [, θ=dflt_θ(x)])

Return $f^{-1}(x \,|\, \theta)$ and $J_{f^{-1}}(x \,| \,\theta)$ where $\theta$ is an array of parameters.

Arguments

x::AbstractArray{T}: arguments to pass to the flow element.

θ::AbstractArray{T}: parameters / conditions (default is dflt_θ).

DensityFlows.forward — Function

forward(f, z [, θ=dflt_θ(z)])

Return $f(z \,| \,\theta)$ and $J_{f}(z \,| \,\theta)$ where $\theta$ is an array of parameters.

Arguments

z::AbstractArray{T}: arguments to pass to the flow element.

θ::AbstractArray{T}: parameters / conditions (default is dflt_θ).

DensityFlows.forward! — Function

forward!(f, z [, θ=dflt_θ(z)])

Replace z by $f(z \,| \,\theta)$ where $\theta$ is an array of parameters.

Arguments

z::AbstractArray{T}: arguments to pass to the flow element.

θ::AbstractArray{T}: parameters / conditions (default is dflt_θ).

Save and load

DensityFlows.save_element — Function

save_element(directory, element; kws...)

Recursively save the FlowElement element's weights in directory.

Setting erase = true force deletes any existing directory with the same name.

DensityFlows.load_element — Function

load_element(directory)

Load any FlowElement saved in directory.

DensityFlows.save_flow — Function

save_flow(directory, flow, erase = false)

Save flow in the folder directory. Clear it before if erase is true.

See @clear_and_save_flow and @save_flow

Macros

DensityFlows.@auto_forward! — Macro

auto_forward!(T)

Automatically define a forward! function for type T from forward if there is no possible optimization to be found in writting a specific forward! function.

DensityFlows.@auto_functor — Macro

auto_functor(T)

Automatically define a functor function for type T calling forward.

DensityFlows.@summary — Macro

auto_functor(element)

Call summarize(element).

DensityFlows.@flow_wrapper — Macro

flow_wrapper(funcs...)

Define the flow version of functions funcs defined otherwise for FlowElement with signature (::FlowElement, AbstractArray{T}, AbstractArray{T}), f(flow::Flow, y, θ) = f(flow.model, y, normalizeinput(θ, flow.metadata.θmin, flow.metadata.θ_max))

DensityFlows.@unconditional_wrapper — Macro

unconditional_wrapper(funcs...)

Define the unconditional version of a function with signature f(::Flow, ::AbstractArray, ::AbstractArray) or f(::FlowElement, ::AbstractArray, ::AbstractArray).

Replace f(obj, y, θ) = f(obj, y, dflt_θ(y))

DensityFlows.@save_as_atomic — Macro

save_as_atomic(T)

Make structure T saved in a single JLD2 file.

DensityFlows.@clear_and_save_flow — Macro

clear_and_save_flow(directory, flow)

Clear any pre-existing folder with name directory and save flow in that folder.

See save_flow and @save_flow

DensityFlows.@save_flow — Macro

save_flow(directory, flow)

Save flow in folder directory.

DensityFlows.@load_flow — Macro

load_flow(directory, flow)

Load flow from folder directory.